Move Valid/Verifying indicators to flat structure
